Fix #14044: Negative string parameters from GS were rendered as zero. (#14049)
String parameters are always stored as uint64_t. Negative values are sign-extended to int64_t and then casted to uint64_t.
The same applies to encoded strings. But ScriptText encoded them as int64_t.

(cherry picked from commit 689f55a0eae476ba245a175fbf6816f24ca26cbd)

# Conflicts:
#	src/saveload/saveload.cpp
#	src/saveload/saveload.h
#	src/script/api/script_text.cpp
#	src/tests/string_func.cpp

+++ src/script/api/script_text.cpp
@@ -316,7 +316,8 @@ void ScriptText::ParamCheck::Encode(std::back_insert_iterator<std::string> &outp
 		void operator()(const SQInteger &value)
 		{
 			Utf8Encode(this->output, SCC_ENCODED_NUMERIC);
-			fmt::format_to(this->output, "{:X}", value);
+			/* Sign-extend the value, then store as unsigned */
+			fmt::format_to(this->output, "{:X}", static_cast<uint64_t>(static_cast<int64_t>(value)));
 		}
 
 		void operator()(const ScriptTextRef &value)
